+++ src/main/java/com/klinker/android/twitter/ui/drawer_activities/lists/ChoosenListActivity.java
@@ -65,7 +65,7 @@ public class ChoosenListActivity extends Activity {
 
     private AsyncListView listView;
 
-    private int listId;
+    private long listId;
     private String listName;
 
     private FullScreenSwipeRefreshLayout mPullToRefreshLayout;
@@ -167,7 +167,7 @@ public class ChoosenListActivity extends Activity {
 
         listName = getIntent().getStringExtra("list_name");
 
-        listId = Integer.parseInt(getIntent().getStringExtra("list_id"));
+        listId = Long.parseLong(getIntent().getStringExtra("list_id"));
         actionBar.setTitle(listName);
 
         getLists();
